pyproject.toml
[project]
name = "tag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "genesis-world",
    "gymnasium>=1.1.1",
    "libigl==2.6.1",
    "matplotlib>=3.10.1",
    "mediapy>=1.2.2",
    "pytest>=8.3.5",
    "rich>=14.0.0",
    "torch>=2.7.0",
    "tqdm>=4.67.1",
    "tyro>=0.9.20",
    "wandb>=0.19.11",
    "jax[cuda12]>=0.5.3",
    "trimesh>=4.6.6",
    "open3d>=0.19.0",
    "tensorboard>=2.19.0",
]

[project.optional-dependencies]
gl = [
    "pyopengl>=3.1.0",
    "pyopengl-accelerate>=3.1.9",
]
debug = [
    "ipdb>=0.13.13",
    "ipython>=9.0.2",
]
dev = [
    "pre-commit>=4.2.0",
]

[tool.setuptools]
packages = ["tag"]

# [tool.uv.sources]
# genesis-lr = { git = "https://github.com/lupinjia/genesis_lr" }
# rsl-rl = { git = "https://github.com/leggedrobotics/rsl_rl", rev = "v1.0.2" }

# [[tool.uv.ind}x]]
# name = "pytorch-cu124"
# url = "https://download.pytorch.org/whl/cu124"
# explicit = true
# [tool.uv.sources]
# torch = [
  # { index = "pytorch-cu124", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
# ]
# torchvision = [
  # { index = "pytorch-cu124", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
# ]

[tool.ruff]
line-length = 120
exclude = ["**/__init__.py"]

[tool.ruff.lint]
select = [
  "E",  # pycodestyle
  "F",  # pyflakes
  "I",  # isort
]
ignore = [
  "F841",  # assigned but unused variable
  "F821",  # undefined name
  "E731",  # do not assign lambda
  "E722",  # do not use bare except
  "E402",  # module import not at top of file
  "E501",  # line too long
  "E741", # ambiguous variable name
]


[tool.ruff.lint.isort]
# force-single-line = true
force-sort-within-sections = true
